limit untrusted contract transactions
This is a change to the split() function to try to take the note regarding only talking to one untrusted contract at a time. I don't know if this is mitigates that threat, but the attempt here is to "wait" for one transaction to complete, before sending the next, by checking to make sure the first transaction correctly sent the appropriate funds. Unless there is a way to wait to receive a confirmation from the first transaction. Or waiting to send the next second transaction in the next block. I am not sure what will best address the refactor request.